week4/requirements.txt
# AI 知识库 V4 — Python 依赖

# --- V2 继承：LLM 客户端 ---
httpx>=0.27.0
python-dotenv>=1.0.0

# --- V2 继承：RSS 源解析 ---
pyyaml>=6.0

# --- V3 继承：LangGraph 工作流引擎 ---
langgraph>=0.2.0
langchain-core>=0.3.0

# --- V4 新增：异步 HTTP 客户端（Telegram / 飞书推送） ---
aiohttp>=3.9.0

# --- 工具链 ---
typing-extensions>=4.9.0
